requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
spacy
geopy
firebase-admin
urllib3==1.26.15
cachecontrol
googlemaps